Updated group_measurement.rb
Debated whether to make a pull request to the README or to the actual code base, but the explanation for why the instance was changed to taken_on vs. created_at seemed to make a lot of sense and seemed more in line with intent of the change than taken_at.

Changed all references and calls in the code to taken_at to taken_on. Feel free to reject and Ill just make the change to the README if that's preferred.